backend/requirements.txt
astroid==2.15.8
black==23.3.0
blinker==1.8.2
click==8.1.8
dill==0.4.0
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
gunicorn==20.1.0
importlib-metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
jinja2==3.1.6
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.6
pylint==2.17.4
python-dotenv==1.0.0
six==1.17.0
sqlalchemy==2.0.41
tomli==2.2.1
tomlkit==0.13.3
typing-extensions==4.13.2
werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
